separated attributes
